// src/app/shared/interfaces/images.interface.ts
export interface CartIcons {
  CIRCLE: string;
  CART: string;
  CIRCLE_UN_AUTH: string;
}

export interface HeartIcons {
  CIRCLE: string;
  HEART: string;
  CIRCLE_UN_AUTH: string;
}

export interface HeaderIcons {
  LOGO: string;
  LOGO_SMALL: string;
  QUESTION: string;
  SEARCH_DESKTOP: string;
  SEARCH_MOBILE: string;
  DOWN_ARROW: string;
  DOWN_ARROW_SMALL: string;
  UP_ARROW: string;
  MENU: string;
  CART: CartIcons;
  HEART: HeartIcons;
}

export interface WinesIcons {
  BOTTLE: string;
  LITTLE_BOTTLE: string;
}

export interface ProductListIcons {
  CANCEL: string;
  MINUS: string;
  PLUS: string;
}

export interface FooterIcons {
  TWITTER: string;
  FACEBOOK: string;
  YOUTUBE: string;
  INSTAGRAM: string;
  GOOGLE: string;
  GROUP_2X: string;
  GROUP: string;
  TWO_MEDALS: string;
  WHITE_UP_ARROW: string;
  WHITE_DOWN_ARROW: string;
  FOOTER_LOGOS: string;
}

export interface FormIcons {
  GOLD_DOWN_ARROW: string;
}